simulator: remove old IPC option for verilator
It is too much overhead to maintain both and
with this PR, we support caching the shared
library.
