+++ src/main/scala/chiseltest/simulator/VerilatorSimulator.scala
@@ -2,10 +2,9 @@
 
 package chiseltest.simulator
 
-import chiseltest.simulator.ipc.{IPCSimulatorContext, VerilatorCppHarnessGenerator}
-import chiseltest.simulator.jna.{JNASimulatorContext, JNAUtils, VerilatorCppJNAHarnessGenerator}
 import firrtl._
-import firrtl.annotations.{JsonProtocol, NoTargetAnnotation}
+import firrtl.annotations._
+import chiseltest.simulator.jna._
 
 case object VerilatorBackendAnnotation extends SimulatorAnnotation {
   override def getSimulator: Simulator = VerilatorSimulator
@@ -22,9 +21,6 @@ case class VerilatorCFlags(flags: Seq[String]) extends VerilatorOption
 
 /** adds flags to the linker in the Makefile generated by verilator */
 case class VerilatorLinkFlags(flags: Seq[String]) extends VerilatorOption
-
-/** tells the backend to use an external process instead of loading a library via JNA */
-private case object VerilatorUseExternalProcess extends VerilatorOption
 
 private object VerilatorSimulator extends Simulator {
   override def name: String = "verilator"
@@ -69,17 +65,8 @@ private object VerilatorSimulator extends Simulator {
     * @param state LoFirrtl circuit + annotations
     */
   override def createContext(state: CircuitState): SimulatorContext = {
-    val shouldCache = Caching.shouldCache(state)
-    val useJNA = !state.annotations.contains(VerilatorUseExternalProcess)
-    if (shouldCache) {
-      if (useJNA) {
-        println("WARN: caching not supported with JNA interface, switching to old IPC interface")
-      }
-      val simName = s"${VerilatorSimulator.name} ${VerilatorSimulator.majorVersion}.${VerilatorSimulator.minorVersion}"
-      Caching.cacheSimulationBin(simName, state, createContextFromScratch(_, useJNA = false), recreateCachedContext)
-    } else {
-      createContextFromScratch(state, useJNA)
-    }
+    val simName = s"${VerilatorSimulator.name} ${VerilatorSimulator.majorVersion}.${VerilatorSimulator.minorVersion}"
+    Caching.cacheSimulationBin(simName, state, createContextFromScratch, recreateCachedContext)
   }
 
   private def recreateCachedContext(state: CircuitState): SimulatorContext = {
@@ -87,7 +74,7 @@ private object VerilatorSimulator extends Simulator {
     val targetDir = Compiler.requireTargetDir(state.annotations)
     val toplevel = TopmoduleInfo(state.circuit)
 
-    val simCmd = Seq("verilated" + "/V" + toplevel.name)
+    val libPath = targetDir / "verilated" / ("V" + toplevel.name)
     // the binary we created communicates using our standard IPC interface
     val coverageAnnos = loadCoverageAnnos(targetDir)
     val coverageFile = targetDir / "coverage.dat"
@@ -95,23 +82,25 @@ private object VerilatorSimulator extends Simulator {
       assert(os.exists(coverageFile), s"Could not find `$coverageFile` file!")
       VerilatorCoverage.loadCoverage(coverageAnnos, coverageFile)
     }
-    new IPCSimulatorContext(simCmd, targetDir, toplevel, VerilatorSimulator, Some(readCoverage))
+
+    val lib = JNAUtils.compileAndLoadJNAClass(libPath)
+    new JNASimulatorContext(lib, targetDir, toplevel, VerilatorSimulator, Some(readCoverage))
   }
 
-  private def createContextFromScratch(state: CircuitState, useJNA: Boolean): SimulatorContext = {
+  private def createContextFromScratch(state: CircuitState): SimulatorContext = {
     // we will create the simulation in the target directory
     val targetDir = Compiler.requireTargetDir(state.annotations)
     val toplevel = TopmoduleInfo(state.circuit)
 
     // Create the header files that verilator needs + a custom harness
     val waveformExt = Simulator.getWavformFormat(state.annotations)
-    val (cppHarness, jnaClassName) = generateHarness(targetDir, toplevel, waveformExt, useJNA)
+    val cppHarness = generateHarness(targetDir, toplevel, waveformExt)
 
     // compile low firrtl to System Verilog for verilator to use
     val verilogState = Compiler.lowFirrtlToSystemVerilog(state, VerilatorCoverage.CoveragePasses)
 
     // turn SystemVerilog into C++ simulation
-    val verilatedDir = runVerilator(toplevel.name, targetDir, cppHarness, state.annotations, useJNA)
+    val verilatedDir = runVerilator(toplevel.name, targetDir, cppHarness, state.annotations)
 
     // patch the coverage cpp provided with verilator only if Verilator is older than 4.202
     // Starting with Verilator 4.202, the whole patch coverage hack is no longer necessary
@@ -121,7 +110,7 @@ private object VerilatorSimulator extends Simulator {
       VerilatorPatchCoverageCpp(verilatedDir, majorVersion, minorVersion)
     }
 
-    val simBin = compileSimulation(topName = toplevel.name, verilatedDir)
+    val libPath = compileSimulation(topName = toplevel.name, verilatedDir)
 
     // the binary we created communicates using our standard IPC interface
     val coverageAnnos = VerilatorCoverage.collectCoverageAnnotations(verilogState.annotations)
@@ -133,15 +122,9 @@ private object VerilatorSimulator extends Simulator {
       assert(os.exists(coverageFile), s"Could not find `$coverageFile` file!")
       VerilatorCoverage.loadCoverage(coverageAnnos, coverageFile)
     }
-    if (useJNA) {
-      // make a copy of the library so that it won't be overwritten
-      val libPath = targetDir / jnaClassName
-      os.copy.over(targetDir / os.RelPath(simBin), to = libPath)
-      val lib = JNAUtils.compileAndLoadJNAClass(libPath)
-      new JNASimulatorContext(lib, targetDir, toplevel, VerilatorSimulator, Some(readCoverage))
-    } else {
-      new IPCSimulatorContext(Seq(simBin), targetDir, toplevel, VerilatorSimulator, Some(readCoverage))
-    }
+
+    val lib = JNAUtils.compileAndLoadJNAClass(libPath)
+    new JNASimulatorContext(lib, targetDir, toplevel, VerilatorSimulator, Some(readCoverage))
   }
 
   private def saveCoverageAnnos(targetDir: os.Path, annos: AnnotationSeq): Unit = {
@@ -152,7 +135,7 @@ private object VerilatorSimulator extends Simulator {
     JsonProtocol.deserialize((targetDir / "coverageAnnotations.json").toIO)
   }
 
-  private def compileSimulation(topName: String, verilatedDir: os.Path): String = {
+  private def compileSimulation(topName: String, verilatedDir: os.Path): os.Path = {
     val target = s"V$topName"
     val cmd = Seq("make", "-C", verilatedDir.toString(), "-j", "-f", s"V$topName.mk", target)
     val ret = os.proc(cmd).call()
@@ -162,7 +145,7 @@ private object VerilatorSimulator extends Simulator {
     )
     val simBinary = verilatedDir / target
     assert(os.exists(simBinary), s"Failed to generate simulation binary: $simBinary")
-    "verilated" + "/" + target
+    simBinary
   }
 
   /** executes verilator in order to generate a C++ simulation */
@@ -170,15 +153,12 @@ private object VerilatorSimulator extends Simulator {
     topName:    String,
     targetDir:  os.Path,
     cppHarness: String,
-    annos:      AnnotationSeq,
-    useJNA:     Boolean
+    annos:      AnnotationSeq
   ): os.Path = {
     val verilatedDir = targetDir / "verilated"
 
     removeOldCode(verilatedDir)
-    val flagAnnos: AnnotationSeq = if (useJNA) {
-      VerilatorLinkFlags(JNAUtils.ldFlags) +: VerilatorCFlags(JNAUtils.ccFlags) +: annos
-    } else { annos }
+    val flagAnnos = VerilatorLinkFlags(JNAUtils.ldFlags) +: VerilatorCFlags(JNAUtils.ccFlags) +: annos
     val flags = generateFlags(topName, verilatedDir, flagAnnos)
     val cmd = List("verilator", "--cc", "--exe", cppHarness) ++ flags ++ List(s"$topName.sv")
     val ret = os.proc(cmd).call(cwd = targetDir)
@@ -244,40 +224,23 @@ private object VerilatorSimulator extends Simulator {
   private def generateHarness(
     targetDir:   os.Path,
     toplevel:    TopmoduleInfo,
-    waveformExt: String,
-    useJNA:      Boolean
-  ): (String, String) = {
+    waveformExt: String
+  ): String = {
     val topName = toplevel.name
 
-    // Create the header files that verilator needs + a custom harness
-    if (!useJNA) {
-      CopyVerilatorHeaderFiles(targetDir)
-    }
+    // create a custom c++ harness
     val cppHarnessFileName = s"${topName}-harness.cpp"
     val vcdFile = targetDir / (s"$topName." + waveformExt)
-    val className = if (useJNA) {
-      val (code, className) = VerilatorCppJNAHarnessGenerator.codeGen(
-        toplevel,
-        vcdFile,
-        targetDir,
-        majorVersion = majorVersion,
-        minorVersion = minorVersion
-      )
-      os.write.over(targetDir / cppHarnessFileName, code)
-      className
-    } else {
-      val code = VerilatorCppHarnessGenerator.codeGen(
-        toplevel,
-        vcdFile,
-        targetDir,
-        majorVersion = majorVersion,
-        minorVersion = minorVersion
-      )
-      os.write.over(targetDir / cppHarnessFileName, code)
-      ""
-    }
+    val code = VerilatorCppJNAHarnessGenerator.codeGen(
+      toplevel,
+      vcdFile,
+      targetDir,
+      majorVersion = majorVersion,
+      minorVersion = minorVersion
+    )
+    os.write.over(targetDir / cppHarnessFileName, code)
 
-    (cppHarnessFileName, className)
+    cppHarnessFileName
   }
 }
 
@@ -368,20 +331,5 @@ private object VerilatorPatchCoverageCpp {
        |}
        |#endif
        |""".stripMargin
-  }
-}
-
-/** Copies the necessary header files used for verilator compilation to the specified destination folder
-  */
-private object CopyVerilatorHeaderFiles {
-  def apply(destinationDirPath: os.Path): Unit = {
-    val files = Seq("sim_api.h", "veri_api.h")
-    val resourcePrefix = "/simulator/"
-
-    files.foreach { name =>
-      val dst = destinationDirPath / name
-      val src = getClass.getResourceAsStream(resourcePrefix + name)
-      os.write.over(target = dst, data = src, createFolders = true)
-    }
   }
 }
