+++ src/test/scala/chiseltest/simulator/Verilator.scala
@@ -2,7 +2,6 @@
 
 package chiseltest.simulator
 
-import firrtl.CircuitState
 import org.scalatest.Tag
 import org.scalatest.flatspec.AnyFlatSpec
 
@@ -27,20 +26,3 @@ class VerilatorSpecificTests extends AnyFlatSpec {
     assert(out.contains("Found Verilator 4"))
   }
 }
-
-object IPCVerilatorSim extends Simulator {
-  override def name = VerilatorSimulator.name + "_ipc" // we need a suffix in order to get different test folders
-  override def isAvailable = VerilatorSimulator.isAvailable
-  override def supportsCoverage = true
-  override def waveformFormats = VerilatorSimulator.waveformFormats
-  override def createContext(state: CircuitState) = {
-    val patchedState = state.copy(annotations = VerilatorUseExternalProcess +: state.annotations)
-    VerilatorSimulator.createContext(patchedState)
-  }
-}
-
-class VerilatorIPCBasicCompliance extends BasicCompliance(IPCVerilatorSim, RequiresVerilator, skipSimRefTest = true)
-class VerilatorIPCStepCompliance extends StepCompliance(IPCVerilatorSim, RequiresVerilator)
-class VerilatorIPCPeekPokeCompliance extends PeekPokeCompliance(IPCVerilatorSim, RequiresVerilator)
-class VerilatorIPCWaveformCompliance extends WaveformCompliance(IPCVerilatorSim, RequiresVerilator)
-class VerilatorIPCCoverageCompliance extends CoverageCompliance(IPCVerilatorSim, RequiresVerilator)
